Site Editor: Add Post Taxonomies panel
