+++ packages/edit-site/src/components/sidebar-edit-mode/page-panels/index.js
@@ -12,7 +12,11 @@ import { humanTimeDiff } from '@wordpress/date';
 import { useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 import { decodeEntities } from '@wordpress/html-entities';
-import { PostLastRevisionPanel, store as editorStore } from '@wordpress/editor';
+import {
+	PostLastRevisionPanel,
+	PostTaxonomiesPanel,
+	store as editorStore,
+} from '@wordpress/editor';
 
 /**
  * Internal dependencies
@@ -95,6 +99,7 @@ export default function PagePanels() {
 				</PanelBody>
 			) }
 			<PostLastRevisionPanel />
+			<PostTaxonomiesPanel />
 		</>
 	);
 }
