+++ packages/edit-site/src/components/sidebar-edit-mode/template-panel/index.js
@@ -3,7 +3,11 @@
  */
 import { useSelect } from '@wordpress/data';
 import { PanelBody } from '@wordpress/components';
-import { PostLastRevisionPanel, store as editorStore } from '@wordpress/editor';
+import {
+	PostTaxonomiesPanel,
+	PostLastRevisionPanel,
+	store as editorStore,
+} from '@wordpress/editor';
 import { store as coreStore } from '@wordpress/core-data';
 import { decodeEntities } from '@wordpress/html-entities';
 import { navigation, symbol } from '@wordpress/icons';
@@ -15,8 +19,6 @@ import { store as editSiteStore } from '../../../store';
 import TemplateActions from './template-actions';
 import TemplateAreas from './template-areas';
 import SidebarCard from '../sidebar-card';
-import PatternCategories from './pattern-categories';
-import { PATTERN_TYPES } from '../../../utils/constants';
 
 const CARD_ICONS = {
 	wp_block: symbol,
@@ -24,29 +26,24 @@ const CARD_ICONS = {
 };
 
 export default function TemplatePanel() {
-	const { title, description, icon, record, postType } = useSelect(
-		( select ) => {
-			const { getEditedPostType, getEditedPostId } =
-				select( editSiteStore );
-			const { getEditedEntityRecord } = select( coreStore );
-			const { __experimentalGetTemplateInfo: getTemplateInfo } =
-				select( editorStore );
+	const { title, description, icon, record } = useSelect( ( select ) => {
+		const { getEditedPostType, getEditedPostId } = select( editSiteStore );
+		const { getEditedEntityRecord } = select( coreStore );
+		const { __experimentalGetTemplateInfo: getTemplateInfo } =
+			select( editorStore );
 
-			const type = getEditedPostType();
-			const postId = getEditedPostId();
-			const _record = getEditedEntityRecord( 'postType', type, postId );
-			const info = getTemplateInfo( _record );
+		const type = getEditedPostType();
+		const postId = getEditedPostId();
+		const _record = getEditedEntityRecord( 'postType', type, postId );
+		const info = getTemplateInfo( _record );
 
-			return {
-				title: info.title,
-				description: info.description,
-				icon: info.icon,
-				record: _record,
-				postType: type,
-			};
-		},
-		[]
-	);
+		return {
+			title: info.title,
+			description: info.description,
+			icon: info.icon,
+			record: _record,
+		};
+	}, [] );
 
 	if ( ! title && ! description ) {
 		return null;
@@ -66,11 +63,7 @@ export default function TemplatePanel() {
 				</SidebarCard>
 			</PanelBody>
 			<PostLastRevisionPanel />
-			<PanelBody>
-				{ postType === PATTERN_TYPES.user && (
-					<PatternCategories post={ record } />
-				) }
-			</PanelBody>
+			<PostTaxonomiesPanel />
 		</>
 	);
 }
